requirements/base.txt
# Core
Django==5.0.1
python-decouple==3.8
django-environ==0.11.2

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pandas
openpyxl==3.1.2
XlsxWriter==3.1.9
python-docx==1.1.0
Pillow
matplotlib==3.8.2

# AI/Agents
smolagents==0.1.0
langchain==0.1.0
openai==1.6.1
tiktoken==0.5.2

# Async Processing
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Frontend
django-htmx==1.17.2
django-widget-tweaks==1.5.0

# Security & Utils
django-cors-headers==4.3.1
django-ratelimit==4.1.0
python-magic==0.4.27
django-cleanup==8.0.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9